+++ web/pwa/src/app/service/feature.service.ts
@@ -32,7 +32,7 @@ export class FeatureService {
             ${this.featureStyle('indicator-fix', isIndicatorFixApplicable())}
             ${this.featureStyle('clipboard-integration', this.clipboardService.isSupported())}
             ${this.featureStyle('native-app', isIOSNative)}
-            ${this.featureStyle('ios-not-native', isIOS && !isIOSNative)}
+            ${this.featureStyle('ios-not-nativee', isIOS && !isIOSNative)}
             ${this.featureStyle('ios-or-safari', isIOS || isSafari)}
             ${this.featureStyle('ios', isIOS)}
         `;